reorganize where the dataset service provider interfaces are kept.
