+++ cdm/core/src/main/java/thredds/client/catalog/tools/CatalogNetcdfFileProvider.java
@@ -5,7 +5,7 @@ import java.util.Formatter;
 import thredds.client.catalog.ServiceType;
 import ucar.nc2.NetcdfFile;
 import ucar.nc2.dataset.DatasetUrl;
-import ucar.nc2.dataset.NetcdfFileProvider;
+import ucar.nc2.dataset.spi.NetcdfFileProvider;
 import ucar.nc2.util.CancelTask;
 
 public class CatalogNetcdfFileProvider implements NetcdfFileProvider {
