+++ cdm/core/src/main/java/ucar/nc2/dataset/spi/NetcdfFileProvider.java
@@ -1,9 +1,13 @@
-package ucar.nc2.dataset;
+package ucar.nc2.dataset.spi;
 
 import java.io.IOException;
 import ucar.nc2.NetcdfFile;
+import ucar.nc2.dataset.DatasetUrl;
 
-/** A Service Provider of NetcdfFile. */
+/**
+ * A Service Provider of NetcdfFile. Typically used by remote protocols (dods, cdmremote)
+ * or anything not reading from a RandomAccessFile.
+ */
 public interface NetcdfFileProvider {
 
   /** The leading protocol string (without a trailing ":"). */
