+++ cdm/core/src/main/java/ucar/nc2/internal/dataset/CoordSystemFactory.java
@@ -16,6 +16,7 @@ import ucar.nc2.Group;
 import ucar.nc2.constants.CDM;
 import ucar.nc2.constants._Coordinate;
 import ucar.nc2.dataset.NetcdfDataset;
+import ucar.nc2.dataset.spi.CoordSystemBuilderFactory;
 import ucar.nc2.internal.dataset.conv.CF1Convention;
 import ucar.nc2.internal.dataset.conv.DefaultConventions;
 import ucar.nc2.internal.ncml.NcMLReaderNew;
