+++ cdm/core/src/main/java/ucar/nc2/internal/dataset/conv/CF1Convention.java
@@ -17,7 +17,7 @@ import ucar.nc2.dataset.NetcdfDataset;
 import ucar.nc2.dataset.TransformType;
 import ucar.nc2.dataset.VariableDS;
 import ucar.nc2.internal.dataset.CoordSystemBuilder;
-import ucar.nc2.internal.dataset.CoordSystemBuilderFactory;
+import ucar.nc2.dataset.spi.CoordSystemBuilderFactory;
 import ucar.nc2.time.CalendarDateUnit;
 import ucar.nc2.units.SimpleUnit;
 import ucar.nc2.util.CancelTask;
