+++ cdm/core/src/main/java/ucar/nc2/ncml/NcmlNetcdfFileProvider.java
@@ -4,7 +4,7 @@ import java.io.IOException;
 import thredds.client.catalog.ServiceType;
 import ucar.nc2.NetcdfFile;
 import ucar.nc2.dataset.DatasetUrl;
-import ucar.nc2.dataset.NetcdfFileProvider;
+import ucar.nc2.dataset.spi.NetcdfFileProvider;
 import ucar.nc2.util.CancelTask;
 
 public class NcmlNetcdfFileProvider implements NetcdfFileProvider {
