+++ dap4/d4cdm/src/main/java/dap4/cdm/nc2/DapNetcdfFileProvider.java
@@ -6,7 +6,7 @@ import org.slf4j.LoggerFactory;
 import thredds.client.catalog.ServiceType;
 import ucar.nc2.NetcdfFile;
 import ucar.nc2.dataset.DatasetUrl;
-import ucar.nc2.dataset.NetcdfFileProvider;
+import ucar.nc2.dataset.spi.NetcdfFileProvider;
 import ucar.nc2.util.CancelTask;
 
 public class DapNetcdfFileProvider implements NetcdfFileProvider {
